CLI_tool/pyproject.toml
[tool.poetry]
name = "cli-tool"
version = "0.1.0"
description = "This is my cli pentesting tool."
authors = ["Roman Shubin <roman.shubin27@gmail.com>"]
readme = "README.md"
packages = [{include = "cli_tool"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.2.1"
requests = "^2.31.0"
rich = "^13.7.0"
PyYAML = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.9.0"
flake8 = "^6.1.0"
mypy = "^1.6.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
pentestctl = "cli_tool.cli:main"